+++ src/Compilers/Lua/Test/Portable/Scoping/FindScopeTests.cs
@@ -1,4 +1,5 @@
 ﻿using Loretta.CodeAnalysis.Lua.Syntax;
+using Loretta.Test.Utilities;
 using Xunit;
 
 namespace Loretta.CodeAnalysis.Lua.UnitTests.Scoping
@@ -31,6 +32,28 @@ namespace Loretta.CodeAnalysis.Lua.UnitTests.Scoping
             var printExpressionScope = script.FindScope(printExpression.Expression);
 
             Assert.Equal(compilationUnitScope, printExpressionScope);
+        }
+
+        [Fact]
+        [Trait("Category", "Script/FindScope")]
+        [WorkItem(106, "https://github.com/LorettaDevs/Loretta/issues/106")]
+        public void FindScope_LocalFunctionIsParsed()
+        {
+            var (tree, script) = ParseScript("local function a() end");
+            var fileScope = script.GetScope(tree.GetRoot());
+
+            Assert.Equal(1, fileScope?.ContainedScopes.Count());
+        }        
+
+        [Fact]
+        [Trait("Category", "Script/FindScope")]
+        [WorkItem(106, "https://github.com/LorettaDevs/Loretta/issues/106")]
+        public void FindScope_AnonymousFunctionIsParsed()
+        {
+            var (tree, script) = ParseScript("(function(Variable) end)()");
+            var fileScope = script.GetScope(tree.GetRoot());
+
+            Assert.Equal(1, fileScope?.ContainedScopes.Count());
         }
     }
 }