+++ CHANGES.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fixed `ContainedScopes` not being populated.
+
 ## v0.2.10
 ### Added
 - Added the following new `SyntaxFactory` overloads to partially restore compatibility with pre-typed-lua era: